requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-session==0.8.0
bcrypt==4.1.2
python-dotenv==1.0.1
requests==2.31.0
reportlab==4.1.0
pypdf==6.7.1
